- recompute random nodes for each iteration
