+++ scripts/compute.py
@@ -1,10 +1,39 @@
 import logging
 import os
-
+import osmnx as ox
 import networkx as nx
 import pandas as pd
 import time
 import openpyxl #Shows as not used but is needed to export .xlsx files
+from shapely import LineString
+import geopandas as gpd
+
+import random_nodes
+
+def get_nodes_geojson(graph, osmid_list):
+    # Create a GeoDataFrame of the nodes in the graph that are in osmid_list
+    nodes_gdf = ox.graph_to_gdfs(graph, edges=False)
+    nodes_gdf = nodes_gdf[nodes_gdf.index.isin(osmid_list)]
+    # Convert the GeoDataFrame to a GeoJSON
+    nodes_geojson = nodes_gdf.to_crs("EPSG:4326").to_json()
+    return nodes_geojson
+
+
+def shortest_path_geojson(G, point1, point2, weight):
+    route = nx.shortest_path(G, point1, point2, weight=weight)
+    nodes = set(route)
+    edges = G.subgraph(nodes)
+    features = gpd.GeoDataFrame(edges.edges(keys=True))
+    features['geometry'] = features.apply(
+        lambda x: LineString([(G.nodes[x[0]]['lon'], G.nodes[x[0]]['lat']), (G.nodes[x[1]]['lon'], G.nodes[x[1]]['lat'])]),
+        axis=1)
+    features['osmid'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('osmid', ''), axis=1)
+    features['name'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('name', ''), axis=1)
+    features['maxspeed'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('fixedmaxspeed', ''), axis=1)
+    features['timetravel'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('traveltimes', ''), axis=1)
+    features['evi_local'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('evi_local', ''), axis=1)
+
+    return features.to_json()
 
 #Need to be duplicated from filter3 because of error "circular import"
 def setup_logger(name, log_file, level=logging.DEBUG):
@@ -19,7 +48,7 @@ def setup_logger(name, log_file, level=logging.DEBUG):
 
     return logger
 
-def compute(graph_proj,rand_nodes,source_node,dest_node,user,sim):
+def compute(graph_proj,graph_not_proj,point1,point2,dist,user,sim):
     # Creation of logger
     os.makedirs("scripts/" + user, exist_ok=True)
     LOG_FILENAME = os.getcwd() + "/scripts/" + user + "/" + sim + "_3.log"
@@ -28,9 +57,19 @@ def compute(graph_proj,rand_nodes,source_node,dest_node,user,sim):
 
     time_start = time.perf_counter()
 
+    source_node = ox.distance.nearest_nodes(graph_not_proj, point1[0], point1[1])
+    dest_node = ox.distance.nearest_nodes(graph_not_proj, point2[0], point2[1])
+
     #TODO Début iteration
     nb_iteration = 3
     for iteration in range(nb_iteration):
+
+        rand_nodes = random_nodes.random_nodes(graph_proj, graph_not_proj, point1[0], point1[1], point2[0], point2[1], user, sim,
+                                               dist)
+        rand_nodes_geojson = get_nodes_geojson(graph_proj, rand_nodes)
+        # with open('rand_nodes.geojson', 'w') as f:
+        #     f.write(rand_nodes_geojson)
+        print(rand_nodes_geojson)
         # base shortest paths
         routes_traveltimes = dict([])
         timetravel_shortest_paths = dict([])
@@ -169,6 +208,11 @@ def compute(graph_proj,rand_nodes,source_node,dest_node,user,sim):
         # df_essential_mw_edges.to_excel("Set_37_essential_mw_edges_test1.xlsx")
 
         nx.set_edge_attributes(graph_proj, evi_local_dict, "evi_local")
+
+        selected_route_geojson = shortest_path_geojson(graph_proj, source_node, dest_node, 'traveltimes')
+        # with open('selected_route.geojson', 'w') as f:
+        #     f.write(selected_route_geojson)
+        print(selected_route_geojson)
     #TODO FIN ITERATION
 
     # calculate computational time
