+++ scripts/filter.py
@@ -24,31 +24,6 @@ def setup_logger(name, log_file, level=logging.DEBUG):
 
     return logger
 
-def get_nodes_geojson(graph, osmid_list):
-    # Create a GeoDataFrame of the nodes in the graph that are in osmid_list
-    nodes_gdf = ox.graph_to_gdfs(graph, edges=False)
-    nodes_gdf = nodes_gdf[nodes_gdf.index.isin(osmid_list)]
-    # Convert the GeoDataFrame to a GeoJSON
-    nodes_geojson = nodes_gdf.to_crs("EPSG:4326").to_json()
-    return nodes_geojson
-
-
-def shortest_path_geojson(G, point1, point2, weight):
-    route = nx.shortest_path(G, point1, point2, weight=weight)
-    nodes = set(route)
-    edges = G.subgraph(nodes)
-    features = gpd.GeoDataFrame(edges.edges(keys=True))
-    features['geometry'] = features.apply(
-        lambda x: LineString([(G.nodes[x[0]]['lon'], G.nodes[x[0]]['lat']), (G.nodes[x[1]]['lon'], G.nodes[x[1]]['lat'])]),
-        axis=1)
-    features['osmid'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('osmid', ''), axis=1)
-    features['name'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('name', ''), axis=1)
-    features['maxspeed'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('fixedmaxspeed', ''), axis=1)
-    features['timetravel'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('traveltimes', ''), axis=1)
-    features['evi_local'] = features.apply(lambda x: G.edges[(x[0], x[1], x[2])].get('evi_local', ''), axis=1)
-
-    return features.to_json()
-
 #################SELECTION OF THE ROADNETWORK#######################################################
 parser = argparse.ArgumentParser()
 # user name
@@ -154,19 +129,7 @@ nx.set_edge_attributes(g, traveltimes, "traveltimes")
 time_elapsed = (time.perf_counter() - time_start)
 logger.info("Filtering time : " + str(time_elapsed))
 logger.info("End of filter")
-rand_nodes = random_nodes.random_nodes(g, g_not_proj,point1[0],point1[1],point2[0],point2[1],user,sim,dist)
-rand_nodes_geojson = get_nodes_geojson(g,rand_nodes)
-# with open('rand_nodes.geojson', 'w') as f:
-#     f.write(rand_nodes_geojson)
-print(rand_nodes_geojson)
 
-source_node = ox.distance.nearest_nodes(g_not_proj, point1[0], point1[1])
-dest_node = ox.distance.nearest_nodes(g_not_proj, point2[0], point2[1])
-
-compute.compute(g, rand_nodes,source_node,dest_node,user,sim)
-selected_route_geojson = shortest_path_geojson(g, source_node, dest_node, 'traveltimes')
-# with open('selected_route.geojson', 'w') as f:
-#     f.write(selected_route_geojson)
-print(selected_route_geojson)
+compute.compute(g, g_not_proj,point1,point2,dist,user,sim)
 
 logger.info("Total time : " + str((time.perf_counter() - time_start)))
